Add new set of tests for checking platform support
Platform tests can be run to ensure browser/platform supports
getUserMedia properly and lay the foundation for mockable tests for the
rest of the code not dependent on platform implementation.

These tests are put in a separate runner as they will be run when
updating/changing browser instead of during implementation of the
GumWrapper library on top.

gumwrapper.platformGetUserMedia is added for finding correct platform
getUserMedia function.